New Layout componentes refactor
