+++ src/components/Layout/Sidebar/index.stories.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+
+import Sidebar from './index'
+
+export default {
+  title: 'Layout/Sidebar',
+  component: Sidebar,
+}
+
+export const Base = (): React.ReactElement => <Sidebar />
